+++ randomGraph/quark.cpp
@@ -16,7 +16,7 @@ void myTask0(Quark * quark)
 void myTask1(Quark * quark)
 {
     unsigned task_id;
-    std::array<uint32_t, 8> *data1;
+    std::array<uint64_t, 8> *data1;
     quark_unpack_args_2( quark, task_id, data1 );
 
     sleep( task_duration );
@@ -25,7 +25,7 @@ void myTask1(Quark * quark)
 void myTask2(Quark * quark)
 {
     unsigned task_id;
-    std::array<uint32_t, 8> *data1, *data2;
+    std::array<uint64_t, 8> *data1, *data2;
     quark_unpack_args_3( quark, task_id, data1, data2 );
 
     sleep( task_duration );
@@ -35,7 +35,7 @@ void myTask2(Quark * quark)
 void myTask3(Quark * quark)
 {
     unsigned task_id;
-    std::array<uint32_t, 8> *data1, *data2, *data3;
+    std::array<uint64_t, 8> *data1, *data2, *data3;
     quark_unpack_args_4( quark, task_id, data1, data2, data3 );
 
     sleep( task_duration );
@@ -46,7 +46,7 @@ void myTask3(Quark * quark)
 void myTask4(Quark * quark)
 {
     unsigned task_id;
-    std::array<uint32_t, 8> *data1, *data2, *data3, *data4;
+    std::array<uint64_t, 8> *data1, *data2, *data3, *data4;
     quark_unpack_args_5( quark, task_id, data1, data2, data3, data4 );
 
     sleep( task_duration );
@@ -58,7 +58,7 @@ void myTask4(Quark * quark)
 void myTask5(Quark * quark)
 {
     unsigned task_id;
-    std::array<uint32_t, 8> *data1, *data2, *data3, *data4, *data5;
+    std::array<uint64_t, 8> *data1, *data2, *data3, *data4, *data5;
     quark_unpack_args_6( quark, task_id, data1, data2, data3, data4, data5 );
 
     sleep( task_duration );
@@ -76,7 +76,7 @@ int main(int argc, char* argv[])
 
     Quark * quark = QUARK_New(n_threads);
 
-    std::vector< std::array<uint32_t, 8> > resources(n_resources);
+    std::vector< std::array<uint64_t, 8> > resources(n_resources);
 
     auto start = high_resolution_clock::now();
 
@@ -89,41 +89,41 @@ int main(int argc, char* argv[])
         case 1:
             QUARK_Insert_Task(quark, myTask1, NULL,
                               sizeof(unsigned), &i, VALUE,
-                              sizeof(std::array<uint32_t,8>), &resources[ access_pattern[i][0] ], INOUT,
+                              sizeof(std::array<uint64_t,8>), &resources[ access_pattern[i][0] ], INOUT,
                               0);
             break;
         case 2:
             QUARK_Insert_Task(quark, myTask2, NULL,
                               sizeof(unsigned), &i, VALUE,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
                               0);
             break;
         case 3:
             QUARK_Insert_Task(quark, myTask3, NULL,
                               sizeof(unsigned), &i, VALUE,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
                               0);
             break;
         case 4:
             QUARK_Insert_Task(quark, myTask4, NULL,
                               sizeof(unsigned), &i, VALUE,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][3] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][3] ], INOUT,
                               0);
             break;
         case 5:
             QUARK_Insert_Task(quark, myTask5, NULL,
                               sizeof(unsigned), &i, VALUE,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][3] ], INOUT,
-                              sizeof(std::array<uint32_t, 8>), &resources[ access_pattern[i][4] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][0] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][1] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][2] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][3] ], INOUT,
+                              sizeof(std::array<uint64_t, 8>), &resources[ access_pattern[i][4] ], INOUT,
                               0);
             break;
         }
